+++ appEditors/AppGeoEditor.py
@@ -1604,8 +1604,9 @@ class FCSelect(DrawTool):
         # self.selected = self.draw_app.selected
 
         # make sure that the cursor text from the FCPath is deleted
-        if self.draw_app.app.plotcanvas.text_cursor.parent:
+        if self.draw_app.app.plotcanvas.text_cursor.parent and self.draw_app.app.use_3d_engine:
             self.draw_app.app.plotcanvas.text_cursor.parent = None
+            self.draw_app.app.plotcanvas.view.camera.zoom_callback = lambda *args: None
 
         # make sure that the Tools tab is removed
         try: